rename: content 도메인의 repository 파일들을 /domain으로 다시 변경
