+++ mopl-api/src/main/java/io/mopl/api/content/domain/ContentRepository.java
@@ -1,10 +1,8 @@
-package io.mopl.api.content.repository;
+package io.mopl.api.content.domain;
 
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import io.mopl.api.content.domain.Content;
 
 @Repository
 public interface ContentRepository extends JpaRepository<Content, UUID>, ContentQueryRepository {}
