+++ mopl-api/src/main/java/io/mopl/api/content/domain/TagRepository.java
@@ -1,4 +1,4 @@
-package io.mopl.api.content.repository;
+package io.mopl.api.content.domain;
 
 import java.util.Collection;
 import java.util.List;
@@ -6,8 +6,6 @@ import java.util.Optional;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import io.mopl.api.content.domain.Tag;
 
 @Repository
 public interface TagRepository extends JpaRepository<Tag, UUID> {
